Upped child limit to 1,000,000 for release coordination purposes.
git-svn-id: https://opencadc.googlecode.com/svn/trunk@1414 728ff76a-78ac-11de-a72b-d90af8dea425

+++ projects/cadcVOS/src/ca/nrc/cadc/vos/server/DatabaseNodePersistence.java
@@ -117,7 +117,7 @@ public abstract class DatabaseNodePersistence implements NodePersistence
 
     protected NodeDAO.NodeSchema nodeSchema;
 
-    protected Integer maxChildLimit = new Integer(1000);
+    protected Integer maxChildLimit = new Integer(1000000);
 
     protected DatabaseNodePersistence(NodeDAO.NodeSchema nodeSchema)
     {
